+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Product Backlog
+* [High] Profile/user.controller.js - create error objects to pass to client
 * [High] Copy Forgot Password functionality from older project
 * [High] Add routes for add to cart buttons in newsletter
 * [Low] Convert login to component (currently loosely coupled controller)